# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: OmniStore AI API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /campaigns/suggestions:
    post:
      operationId: getCampaignSuggestions
      tags: [campaigns]
      summary: AI-generated Hinglish campaign message suggestions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CampaignSuggestionsBody"
      responses:
        "200":
          description: Suggestions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CampaignSuggestionsResult"
...
  /appointments:
    get:
      operationId: listAppointments
      tags: [appointments]
      parameters:
        - name: date
          in: query
          schema:
            type: string
            description: ISO date string for filtering
      responses:
        "200":
          description: Appointments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Appointment"
    post:
      operationId: createAppointment
      tags: [appointments]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAppointmentBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
  /appointments/{id}:
    patch:
      operationId: updateAppointment
      tags: [appointments]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAppointmentBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
    delete:
      operationId: deleteAppointment
      tags: [appointments]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /inventory:
    get:
      operationId: listInventory
      tags: [inventory]
      responses:
        "200":
          description: Inventory
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InventoryItem"
    post:
      operationId: createInventoryItem
      tags: [inventory]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInventoryBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
  /inventory/{id}:
    patch:
      operationId: updateInventoryItem
      tags: [inventory]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateInventoryBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
...
components:
  schemas:
...
    CampaignSuggestionsBody:
      type: object
      properties:
        deadZoneLabel:
          type: string
        discountPercent:
          type: integer
        targetTier:
          type: string
        service:
          type: string
      required: [deadZoneLabel, discountPercent, targetTier, service]
    CampaignSuggestionsResult:
      type: object
      properties:
        suggestions:
          type: array
          items:
            type: string
      required: [suggestions]
...
    Appointment:
      type: object
      properties:
        id:
          type: integer
        customerId:
          type: integer
          nullable: true
        customerName:
          type: string
          nullable: true
        customerPhone:
          type: string
          nullable: true
        staffId:
          type: integer
          nullable: true
        staffName:
          type: string
          nullable: true
        service:
          type: string
        notes:
          type: string
          nullable: true
        status:
          type: string
          enum: [scheduled, completed, cancelled, no_show]
        scheduledAt:
          type: string
          format: date-time
        durationMinutes:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, service, status, scheduledAt, durationMinutes, createdAt]
    CreateAppointmentBody:
      type: object
      properties:
        customerId:
          type: integer
          nullable: true
        staffId:
          type: integer
          nullable: true
        service:
          type: string
        notes:
          type: string
          nullable: true
        scheduledAt:
          type: string
          format: date-time
        durationMinutes:
          type: integer
      required: [service, scheduledAt]
    UpdateAppointmentBody:
      type: object
      properties:
        status:
          type: string
          enum: [scheduled, completed, cancelled, no_show]
        notes:
          type: string
          nullable: true
        scheduledAt:
          type: string
          format: date-time
          nullable: true
...
    InventoryItem:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
        price:
          type: number
        costPrice:
          type: number
        stock:
          type: integer
        lowStockThreshold:
          type: integer
        isService:
          type: boolean
        isActive:
          type: boolean
        description:
          type: string
          nullable: true
        isLowStock:
          type: boolean
        margin:
          type: number
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - category
        - price
        - costPrice
        - stock
        - lowStockThreshold
        - isService
        - isActive
        - isLowStock
        - margin
        - createdAt
    CreateInventoryBody:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
        price:
          type: number
        costPrice:
          type: number
        stock:
          type: integer
        lowStockThreshold:
          type: integer
        isService:
          type: boolean
        description:
          type: string
          nullable: true
      required: [name, category, price, costPrice]
    UpdateInventoryBody:
      type: object
      properties:
        price:
          type: number
          nullable: true
        stock:
          type: integer
          nullable: true
        isActive:
          type: boolean
          nullable: true
...